requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.20.0
oauthlib==3.1.0
pika==1.1.0
protobuf==3.13.0
pylint==2.6.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rope==0.17.0
six==1.15.0
toml==0.10.1
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
